fix(jobs): adjust group/namespace handling
Change the jobs parsing to properly parse jobs into namespaces and
adjust the routing and view logic to handle namespaces and jobs
properly, so that the routes don't collide. This allows jobs and
namespaces to have the same id without being collapsed into one object.
An array of jobs like the following would be parsed into a job named
"foo" and a namespace "foo" which will include the "foo.bar" job.

```
[
  { id: "foo" },
  { id: "foo.bar" }
]
```

```
[
   { id: "foo" },
   { id: "foo",
      items: [
        { id: "foo.bar" }
      ]
   }
]
```

These changes addresses issues with jobs that were accidentally merged
with namespaces.

BREAKING CHANGE: The `MetronomeUtil.parseJobs` output as well as the job
routing has changed to address issues with the job/namespace
representation.

Implementations relying on `MetronomeUtil.parseJobs` need to be adjusted
to handle namespace and jobs with the same id.

To migrate links and bookmarks to job details and namespaces please
adjust the URLs as illustrated below:

Before
* Job Detail: `/jobs/${jobId}`
* Overview: `/jobs`
* Namespace: `/jobs/${namespaceId}`

After
* Job Detail:: /jobs/detail/${jobId}
* Overview: `/jobs/overview`
* Namespace: /jobs/overview/${namespaceId}

Closes DCOS_OSS-1033